requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
rank-bm25>=0.2.2
pyyaml>=6.0
